moved HTML generator into its own file
